Ponderada8/requirements.txt
openai-whisper
deep-translator
openai==0.28